defines in the shader system

+++ src/editor/entity_template_system.cpp
@@ -35,7 +35,7 @@ namespace Lumix
 						: m_entity_system(entity_system)
 						, m_template_name_hash(crc32(template_name))
 						, m_position(position)
-						, m_rotation(Vec3(0, 1, 0), Math::degreesToRadians(rand() % 360))
+						, m_rotation(Vec3(0, 1, 0), Math::degreesToRadians((float)(rand() % 360)))
 					{
 					}
